Backend: Warn for all failed backend loads.
This ensures that if every backend cannot be loaded, the user will be
warned. I previously thought to remove the line that hardcodes loading
the dri GBM backend, but the problem with this is that for some if not
all Mesa drivers, the backend will be the driver name, like nouveau, not
dri, so the proper backend would never get loaded. The hardcoded loading
of dri is a last resort, but it makes the warnings harder to parse for
the reader. Gonna try to find a way to make it easier to read and
understand how to proceed.